MIDI CC record when in synth controller mode will not play CC while CC track record is on, to prevent a feedback loop with a fast MIDI bus.

+++ OCT_OS/_OCT_Player/play__master.h
@@ -210,6 +210,7 @@ void PLAYER_dispatch( unsigned char in_G_TTC_abs_value ) {
 		if (G_delay < 16) return;
 	}
 	#endif
+
 	// Play MIDI queue elements which are due at current timestamp
 	play_MIDI_queue( G_MIDI_timestamp );
 
@@ -296,7 +297,7 @@ void PLAYER_dispatch( unsigned char in_G_TTC_abs_value ) {
 
 			// drum machine scene change
 			if ( GRID_p_set_note_presel != 255) {
-				MIDI_send(	MIDI_NOTE, GRID_p_set_midi_ch, GRID_p_set_note_offsets[GRID_p_set_note_presel], 127 );
+				MIDI_send(	MIDI_NOTE, GRID_p_set_midi_ch, GRID_p_set_note_offsets[(int) GRID_p_set_note_presel], 127 );
 				GRID_p_set_note_presel = 255;
 			}
 		}
